+++ Makefile
@@ -34,7 +34,7 @@ starttestcontainer:
 		--env-file "${HOME}/.pingidentity/config" \
 		-v $$(pwd)/server-profiles/shared-profile:/opt/in \
 		-v $$(pwd)/server-profiles/12.2/data.json:/opt/in/instance/bulk-config/data.json \
-		pingidentity/pingfederate:latest
+		pingidentity/pingfederate:12.2.0-latest
 # Wait for the instance to become ready
 	sleep 1
 	duration=0
